refactor: remove types from files and create imports from types file
