+++ src/components/features/MainPage/index.tsx
@@ -2,11 +2,13 @@ import Navigation from '../../Navigation';
 import { ErrorButton } from '../../error/ErrorButton';
 import { Outlet } from 'react-router';
 
+const mainHeading = 'Rick&Morty React App';
+
 export default function MainPage() {
   return (
     <>
       <header className="header">
-        <h1>Rick&Morty React App</h1>
+        <h1>{mainHeading}</h1>
         <Navigation />
       </header>
       <main className="main">
